backend/shared/requirements.txt
annotated-doc==0.0.2
annotated-types==0.7.0
anyio==4.11.0
cffi==2.0.0
click==8.3.0
cryptography==46.0.3
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.120.0
h11==0.16.0
idna==3.11
passlib==1.7.4
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
PyMySQL==1.1.2
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
httpx==0.28.1
pyJWT==2.10.1
sqlmodel==0.0.27